+++ packages/person/src/index.ts
@@ -1,3 +1,4 @@
-export * from './person-avatar';
-export * from './person-provider';
+export { default as PersonAvatarElement, AvatarSize, PersonAvatarElementProps } from './person-avatar';
+export { default as PersonCardElement } from './person-card';
+export { default as PersonProviderElement, PersonResolver } from './person-provider';
 export * from './types';