+++ packages/person/src/person-card/types.ts
@@ -1,0 +1,5 @@
+import { AvatarSize } from '@equinor/fusion-wc-avatar';
+
+export type PersonCardElementProps = {
+  size?: AvatarSize;
+};